Update post dates and cleanup code block
Updated the publication dates for two blog posts and removed a redundant newline from a code block in the second post for better readability.
This change updates the "Taylor Series approximations of ODEs" and "Modelling Drug Use in Communities" blog posts with their new publication dates in 2023.
Additionally, the commit includes the removal of an unnecessary newline character from the R Markdown code block for the drug use model post, which tidied up the report's code presentation without affecting functionality. It is part of general maintenance to keep the content up-to-date and the code clean.