Add Part 2 of Taylor Series Blog Post
Extended the Taylor Series blog series with Part 2, including Python code and a simulated example comparing methods.
Euler Method and an improved version using Taylor series were described and implemented in Python.
The new method's stepping matrix was contrasted with the traditional Euler Method.
A simulation function was also provided to demonstrate the methods' differences.
Finally, an example comparing the true solution of a projectile's path to both numerical solutions was illustrated with a plot.