Enhancing Numerical ODE Solutions with the Taylor Series (#16)
* Update Taylor series equation in post-1

Correct the order of terms in the Taylor series matrix.

The Taylor series post had an incorrect term order in the matrix.
This update corrects the order by placing `y(x)` at the beginning and
removing an incorrect term previously placed there.

* Add Taylor Series library concept to Post 2

Implemented the Taylor Series improvement over the Euler method in Python.

The post now details the creation of a Python library based on the Taylor Series. Each component and mathematical equation is explained. The improvement over the Euler method is showcased through examples. The post is intended to guide users on how to apply the new method.

* Improve matrix representation in post

Enhanced the visual layout of the Y' matrix.
The matrix is now more readable with values for y(x+h).
The shift aims for clarity and better understanding.

* Add Part 2 of Taylor Series Blog Post

Extended the Taylor Series blog series with Part 2, including Python code and a simulated example comparing methods.
Euler Method and an improved version using Taylor series were described and implemented in Python.
The new method's stepping matrix was contrasted with the traditional Euler Method.
A simulation function was also provided to demonstrate the methods' differences.
Finally, an example comparing the true solution of a projectile's path to both numerical solutions was illustrated with a plot.

* site build