Add new dev container blog post
Added a post about improving Euler Method using Taylor Series.

Alfie Chadwick authored a new blog post for the dev container series.
The post is titled "Using Taylor Series to Improve the Euler Method: Part 1".
This post has been dated for the 18th of December, 2023.
The last modified date is set to update automatically with the system date.
Tags include Devops, Docker, and Vim, indicating the post's content.
The output is set to generate an HTML page with a table of contents.