Update Taylor Series blog post titles and create new entry
Updated post titles and added new Taylor Series post to blog.

This commit includes updates to the Taylor series blog posts. The titles of the existing posts were modified to remove the "Part" designation and become more descriptive. Additionally, a new post titled "Making my ODE solver solve ODEs" was added to the collection, complete with updated metadata, md5sums, and references in the site's indexing and listing files.

The aim was to provide a more intuitive understanding for readers on the content and progression of the series, along with keeping the blog's infrastructure up-to-date with the latest changes.