Update post references and modification dates
This commit updates the references to previous posts in the Taylor Series entries and adjusts the modification dates to reflect the new changes.

Modification dates for relevant HTML documents have been set to "2024-01-13".
In-text and link references to previous posts have been corrected.
The MD5 checksums for posts 2 and 3 of the Taylor Series have been updated, likely due to content changes reflected in the modification dates.

This ensures that all internal references are accurate and that the latest version of the content is correctly represented in checksums and metadata.