Update new post with visualizations tag
Added new post 'Predicting the 2024 Hottest 100' with updated tags and date.
The new post, authored by Alfie Chadwick, is dated '2024-01-26'.
Added 'Visualizations' to the list of tags for the post.
Generated HTML and associated files for the new post.
Added md5sum for the post's R Markdown file to track changes.
Adjusted paths in the public folder reflecting the new post date.
Included the post under the Music and Stats tags as well.