Update data gathering scripts and fix post title
Added scripts for extracting data and fixed the 2024 mislabel.

Added a Python script for scraping historical Hottest 100 data and saving it as CSVs.
Included multiprocessing to handle data retrieval for different years concurrently.
Corrected the year in the post title from 2024 to 2023.
Added a requirements file listing necessary Python packages.