Refactor code and enhance prediction comparison
Restructured song truncation logic and streamlined prediction analysis.

The code changes include a refactoring of the way song titles are truncated in the R script and a streamlined comparison between the prediction performance of "my_100" and "100 Warm Tunas". The truncation logic for song titles has been moved from an if-statement within a loop to a more concise mutate call that applies the same logic. Additionally, description of the prediction comparison has been edited for clarity, and the results analysis has been extended with a new R code block that calculates the average deviation of predictions, showing a quantitative measure of each method's accuracy. Furthermore, the author speculates on potential improvements for future predictions, such as addressing the name joining issue and considering genre and artist nationality in the predictive model.