Added Taylor Swift song-list and bead dictionary
This commit introduces a list of Taylor Swift's songs and a dictionary for beads with representative numeric values mappings for every alphabet letter.
The song-list will be used to find the song title that matches the total beads count.
The bead dictionary translates each bead to a certain number, where each letter of the alphabet is represented as a number from 33 ('a') to 58 ('z').
The aim is to maximize the number of friendship bracelets for a Taylor Swift concert.