Update on ODE approximation explanation
Reworded parts of the post on approximating ODEs, specifically concerning the conversion of ODEs into matrices and the definition of truncation error.

The paragraphs regarding the example ODEs, the matrices that represent them, and the variability in their results based on the defined matrix have been redrafted for clarity.

The observation on truncation error and its introduction in the Euler method is now spared from having been 'glossed over'. The error, represented as the difference between the true and the predicted values, has been succinctly but thoroughly explained.