ODE solver post  (#22)
* Title: Add `post-4.rmd` to discuss ODE solver optimization

A new post was added discussing how different transformation matrices affect the accuracy of the Euler method when approximating solutions to ordinary differential equations (ODEs).
The post provides a more detailed look at the truncation error when using different transformation matrices and provides a mathematical rationale for choosing the best matrix to minimize this error. The mathematical explanation is accompanied by Python code that calculates and visually portrays these errors. The code is organized in functions and a class, and the Matplotlib library is used to create illustrative plots. The explanation utilizes Taylor Series, Big-O notation, and vector magnitudes to make clear which matrix results in the smallest error.

* Update on ODE approximation explanation

Reworded parts of the post on approximating ODEs, specifically concerning the conversion of ODEs into matrices and the definition of truncation error.

The paragraphs regarding the example ODEs, the matrices that represent them, and the variability in their results based on the defined matrix have been redrafted for clarity.

The observation on truncation error and its introduction in the Euler method is now spared from having been 'glossed over'. The error, represented as the difference between the true and the predicted values, has been succinctly but thoroughly explained.

* ODE solver post