Corrected spelling and grammar in post-4.rmd
In the file `post-4.rmd`, several edits were made.

Incorrect spellings of 'equally' and 'surely' were amended.

Additionally, the word 'Optimising' in the title was changed to the American English spelling 'Optimizing'.

A punctuation error was also corrected, in the form of an added comma after the word 'However'.

These changes will improve the readability of the text.